// smartcoin-frontend/src/app/api/model/currencySubtotal.ts
/**
 * Smartcoin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Suma de los saldos actuales de las cuentas de una moneda.
 */
export interface CurrencySubtotal { 
    currency: CurrencySubtotal.CurrencyEnum;
    balance: number;
}
export namespace CurrencySubtotal {
    export const CurrencyEnum = {
        Ars: 'ARS',
        Usd: 'USD',
    } as const;
    export type CurrencyEnum = typeof CurrencyEnum[keyof typeof CurrencyEnum];
}